Move domain card ui state to view model
